+++ azclishell/az_completer.py
@@ -112,13 +112,12 @@ class AzCompleter(Completer):
 
     def validate_completion(self, param, words, text_before_cursor, double=True):
         """ validates that a param should be completed """
-        return param.lower().startswith(words.lower()) and \
-                    param.lower() != words.lower() and\
-                        param not in text_before_cursor.split()\
-                            and not text_before_cursor[-1].isspace() and\
-                                (not (double and param in self.same_param_doubles) or\
-                                    self.same_param_doubles[param] not in\
-                                        text_before_cursor.split())
+        return param.lower().startswith(words.lower()) and param.lower() != words.lower() and\
+                    param not in text_before_cursor.split() and not \
+                        text_before_cursor[-1].isspace() and\
+                            (not (double and param in self.same_param_doubles) or
+                             self.same_param_doubles[param] not in
+                             text_before_cursor.split())
 
     def get_completions(self, document, complete_event):
         text = document.text_before_cursor
@@ -183,10 +182,8 @@ class AzCompleter(Completer):
                 # this try catches which format it is
                 if self.cmdtab[self.curr_command].arguments[arg_name].completer:
                     try:
-                        for comp in self.cmdtab[self.curr_command].\
-                                arguments[arg_name].completer(
-                                        prefix=prefix, action=None,
-                                        parser=None, parsed_args=parse_args):
+                        for comp in self.cmdtab[self.curr_command].arguments[arg_name].completer(
+                                prefix=prefix, action=None, parser=None, parsed_args=parse_args):
 
                             for comp in gen_dyn_completion(
                                     comp, started_param, prefix, text):
@@ -321,4 +318,4 @@ class AzCompleter(Completer):
     def has_description(self, param):
         """ if a parameter has a description """
         return param in self.param_description.keys() and \
-               not self.param_description[param].isspace()
+                  not self.param_description[param].isspace()
